refactor: Extract Features.ListFeatures into proper file
